+++ CHANGELOG.md
@@ -12,7 +12,6 @@
 * Updated Dart SDK version (For Initial Release)
 
 
-
 ## 1.1.1
 
 * Changed supported platforms
@@ -25,3 +24,12 @@
 ## 1.3.0
 
 * Added Windows to the pubspec platforms listing
+
+
+## 1.4.0
+
+* Re-generated the MacOS and Windows generated bindings on native machines
+
+## 2.0.0
+
+* Moved library calls to isolate to prevent blocking calls. Dropped incomplete MacOS and Windows support